+++ packages/stf/src/lib/data-engine.ts
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from 'react';
-import { objectGet, objectSet, stringifyFieldKey } from './utils';
+import { fieldKeyStartsWith, objectGet, objectSet, stringifyFieldKey } from './utils';
 import type { FieldKey } from './types';
 import { Stf, useDataEngine } from './stf';
 
@@ -15,24 +15,39 @@ export interface DataEngineListener {
    */
   field?: FieldKey;
   /**
-   * when field value is changed
+   * when field value is changed.
    */
   onUpdate?: (key: FieldKey, ctx: OnUpdateContext) => void;
   /**
-   * when `init(field)` is called
+   * fired on `init(field)`.
    */
-  onInit?: (key: FieldKey) => void;
+  onInit?: (key: FieldKey, ctx: OnInitContext) => void;
   /**
-   * when `delete(field)` is called
+   * fired on `delete(field)`.
+   *
+   * when `field` is specified, this is also fired when parent is deleted.
    */
-  onDelete?: (key: FieldKey) => void;
+  onDelete?: (key: FieldKey, ctx: OnDeleteContext) => void;
 }
 
-interface OnUpdateContext {
+export interface OnUpdateContext {
   /**
    * An update is swallow if the change doesn't affect the values of children fields.
    */
   swallow: boolean;
+
+  /** custom data */
+  custom?: Record<string, unknown>;
+}
+
+export interface OnDeleteContext {
+  /** custom data */
+  custom?: Record<string, unknown>;
+}
+
+export interface OnInitContext {
+  /** custom data */
+  custom?: Record<string, unknown>;
 }
 
 class ListenerManager {
@@ -67,29 +82,35 @@ class ListenerManager {
       if (set) for (const v of set) v.onUpdate?.(field, ctx);
     } else {
       for (const [k, listeners] of this.indexed.entries()) {
-        if (updatedKey.length !== 0 && k !== updatedKey && !k.startsWith(updatedKey + '.'))
-          continue;
+        if (!fieldKeyStartsWith(k, updatedKey)) continue;
+
         for (const v of listeners) v.onUpdate?.(field, ctx);
       }
     }
   }
 
-  onInit(field: FieldKey) {
-    for (const v of this.unindexed) v.onInit?.(field);
+  onInit(field: FieldKey, ctx: OnInitContext) {
+    for (const v of this.unindexed) v.onInit?.(field, ctx);
     const set = this.indexed.get(stringifyFieldKey(field));
-    if (set) for (const v of set) v.onInit?.(field);
+    if (set) for (const v of set) v.onInit?.(field, ctx);
   }
 
-  onDelete(field: FieldKey) {
-    for (const v of this.unindexed) v.onDelete?.(field);
-    const set = this.indexed.get(stringifyFieldKey(field));
-    if (set) for (const v of set) v.onDelete?.(field);
+  onDelete(field: FieldKey, ctx: OnDeleteContext) {
+    const fieldKey = stringifyFieldKey(field);
+    for (const v of this.unindexed) v.onDelete?.(field, ctx);
+
+    for (const [k, listeners] of this.indexed.entries()) {
+      if (!fieldKeyStartsWith(k, fieldKey)) continue;
+
+      for (const v of listeners) v.onDelete?.(field, ctx);
+    }
   }
 }
 
 export class DataEngine {
   private data: NonNullable<object>;
-  private attachedDataMap = new Map<string, unknown>();
+  /** namespace -> engine */
+  private namespaces = new Map<string, DataEngine>();
   private readonly listeners = new ListenerManager();
 
   constructor(defaultValues: DefaultValue<NonNullable<object>> = {}) {
@@ -114,10 +135,11 @@ export class DataEngine {
    * @param defaultValue the initial value, the field is also created for `undefined`
    * @returns the value of initialized field, or the current value of field if already initialized
    */
-  init(key: FieldKey, defaultValue?: DefaultValue): unknown {
+  init(key: FieldKey, defaultValue?: DefaultValue, ctx: OnInitContext = {}): unknown {
     if (key.length === 0) return this.data;
     let cur = this.data as Record<string, unknown>;
     const currentKey: FieldKey = [];
+    const parentUpdateCtx = { swallow: true, ...ctx };
 
     for (let i = 0; i < key.length; i++) {
       const propKey = key[i];
@@ -127,8 +149,8 @@ export class DataEngine {
         if (propValue !== undefined) return propValue;
         cur[propKey] = getDefaultValue(defaultValue);
 
-        this.listeners.onUpdate(currentKey, { swallow: true });
-        this.listeners.onInit(key);
+        this.listeners.onUpdate(currentKey, parentUpdateCtx);
+        this.listeners.onInit(key, ctx);
         return cur[propKey];
       } else if (typeof propValue === 'object' && propValue !== null) {
         cur = propValue as Record<string, unknown>;
@@ -139,28 +161,31 @@ export class DataEngine {
           );
 
         cur = cur[propKey] = {};
-        this.listeners.onUpdate(currentKey, { swallow: true });
+        this.listeners.onUpdate(currentKey, parentUpdateCtx);
       }
       currentKey.push(propKey);
     }
   }
 
-  delete(key: FieldKey): unknown | undefined {
+  delete(key: FieldKey, ctx: OnDeleteContext = {}): unknown | undefined {
     if (key.length === 0) return;
     const parentKey = key.slice(0, -1);
     const prop = key[key.length - 1];
     const parent = this.get(parentKey);
 
     if (Array.isArray(parent) && typeof prop === 'number') {
-      const [deleted] = parent.splice(prop, 1);
-      this.listeners.onUpdate(parentKey, { swallow: false });
-      this.listeners.onDelete(key);
-      return deleted;
+      const deleted: unknown[] = parent.splice(prop, 1);
+      if (deleted.length === 0) return;
+
+      this.listeners.onDelete(key, ctx);
+      // it will change children's field value when removed at middle
+      this.listeners.onUpdate(parentKey, { swallow: false, ...ctx });
+      return deleted[0];
     } else if (typeof parent === 'object' && parent !== null) {
       const temp = (parent as Record<string, unknown>)[prop];
       delete parent[prop as never];
-      this.listeners.onUpdate(parentKey, { swallow: true });
-      this.listeners.onDelete(key);
+      this.listeners.onDelete(key, ctx);
+      this.listeners.onUpdate(parentKey, { swallow: true, ...ctx });
       return temp;
     }
   }
@@ -173,40 +198,31 @@ export class DataEngine {
    * update the value of field if it exists
    * @returns if the field is updated
    */
-  update(key: FieldKey, value: unknown): boolean {
+  update(key: FieldKey, value: unknown, ctx?: Partial<OnUpdateContext>): boolean {
     try {
       this.data = objectSet(this.data, key, value) as NonNullable<object>;
-      this.listeners.onUpdate(key, { swallow: false });
+      this.listeners.onUpdate(key, { swallow: false, ...ctx });
       return true;
     } catch {
       return false;
     }
   }
 
-  attachedData<T>(namespace: string) {
-    return {
-      get: (field: FieldKey): T | undefined => {
-        return this.attachedDataMap.get(`${namespace}:${stringifyFieldKey(field)}`) as
-          | T
-          | undefined;
-      },
-      set: (field: FieldKey, value: T) => {
-        this.attachedDataMap.set(`${namespace}:${stringifyFieldKey(field)}`, value);
-      },
-      delete: (field?: FieldKey) => {
-        if (field) {
-          this.attachedDataMap.delete(`${namespace}:${stringifyFieldKey(field)}`);
-        } else {
-          for (const key of this.attachedDataMap.keys()) {
-            if (key.startsWith(`${namespace}:`)) this.attachedDataMap.delete(key);
-          }
-        }
-      },
-    };
+  /**
+   * create an isolated data engine
+   */
+  namespace(namespace: string, initialValue?: DefaultValue<NonNullable<object>>) {
+    let child = this.namespaces.get(namespace);
+    if (!child) {
+      child = new DataEngine(initialValue);
+      this.namespaces.set(namespace, child);
+    }
+
+    return child;
   }
 
   reset(data: Record<string, unknown>) {
-    this.attachedDataMap.clear();
+    this.namespaces.clear();
     this.update([], data);
   }
 }
@@ -214,7 +230,7 @@ export class DataEngine {
 export function useFieldValue<V = unknown>(
   key: FieldKey,
   options: {
-    stf?: Stf;
+    stf?: Stf | DataEngine;
     defaultValue?: DefaultValue;
 
     /**
@@ -233,20 +249,25 @@ export function useFieldValue<V = unknown>(
 
   useListener({
     field: key,
+    stf: options.stf,
+    onInit() {
+      const computed = compute(engine.get(key));
+      setValue((prev) => (isChanged(prev, computed) ? computed : prev));
+    },
     onUpdate() {
       const computed = compute(engine.get(key));
-      if (isChanged(value, computed)) setValue(computed);
+      setValue((prev) => (isChanged(prev, computed) ? computed : prev));
     },
     onDelete() {
       const computed = compute(undefined);
-      if (isChanged(value, computed)) setValue(computed);
+      setValue((prev) => (isChanged(prev, computed) ? computed : prev));
     },
   });
 
   return [value, (newValue: unknown) => engine.update(key, newValue)] as const;
 }
 
-export function useListener(listener: DataEngineListener & { stf?: Stf }) {
+export function useListener(listener: DataEngineListener & { stf?: Stf | DataEngine }) {
   const engine = useDataEngine(listener.stf);
   const listenerRef = useRef(listener);
   listenerRef.current = listener;